Reorganzies s3 service into bin and lib files for testing purposes. Adds tests, updates readme, sos metadata, and toml files. Fixes a bug in s3-helloworld.
