+++ .doc_gen/metadata/s3_metadata.yaml
@@ -89,7 +89,7 @@ s3_CreateBucket:
           excerpts:
             - description:
               snippet_tags:
-                - s3.rust.create-bucket
+                - rust.example_code.s3.basics.create_bucket
     Go:
       versions:
         - sdk_version: 2
@@ -187,6 +187,14 @@ s3_CopyObject:
             - description: Copy an object and add server-side encryption to the destination object.
               snippet_tags:
                 - s3.ruby.copy_object_encrypt_copy.rb
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/s3
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.example_code.s3.basics.copy_object
   services:
     s3: {CopyObject}
 s3_DeleteBucketPolicy:
@@ -402,7 +410,7 @@ s3_DeleteObjects:
           excerpts:
             - description:
               snippet_tags:
-                - s3.rust.delete-objects
+                - rust.example_code.s3.basics.create_bucket
   services:
     s3: {DeleteObjects}
 s3_DeleteBucketWebsite:
@@ -668,6 +676,14 @@ s3_GetObject:
             - description: Get an object and report its server-side encryption state.
               snippet_tags:
                 - s3.ruby.determine_object_encryption_state.rb
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/s3
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.example_code.s3.basics.get_object
   services:
     s3: {GetObject}
 s3_GetBucketLifecycleConfiguration:
@@ -816,7 +832,7 @@ s3_ListObjects:
           excerpts:
             - description:
               snippet_tags:
-                - s3.rust.list-objects
+                - rust.example_code.s3.basics.list_objects
     Go:
       versions:
         - sdk_version: 2
@@ -940,7 +956,7 @@ s3_PutObject:
           excerpts:
             - description:
               snippet_tags:
-                - s3.rust.s3-helloworld
+                - rust.example_code.s3.basics.put_object
     Go:
       versions:
         - sdk_version: 2
@@ -1074,6 +1090,14 @@ s3_DeleteBucket:
             - description:
               snippet_tags:
                 - ruby.example_code.s3.DeleteBucket
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/s3
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.example_code.s3.basics.delete_bucket
   services:
     s3: {DeleteBucket}
 s3_GetBucketCors:
@@ -1739,5 +1763,13 @@ s3_Scenario_GettingStarted:
             - description:
               snippet_tags:
                 - ruby.example_code.s3.Scenario_GettingStarted
+    Rust:
+      versions:
+        - sdk_version: 1
+          github: rust_dev_preview/s3
+          excerpts:
+            - description:
+              snippet_tags:
+                - rust.example_code.s3.basics.rust.example_code.s3.scenario_getting_started
   services:
     s3: {CreateBucket, PutObject, GetObject, CopyObject, ListObjects, DeleteObjects, DeleteBucket}
