+++ rust_dev_preview/s3/src/bin/s3-getting-started.rs
@@ -3,16 +3,17 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 
+// snippet-start:[rust.example_code.s3.scenario_getting_started.bin]
+
+// snippet-start:[rust.example_code.s3.basics.imports]
+
 use aws_config::meta::region::RegionProviderChain;
-use aws_sdk_s3::model::{
-    BucketLocationConstraint, CreateBucketConfiguration, Delete, ObjectIdentifier,
-};
-use aws_sdk_s3::output::ListObjectsV2Output;
-use aws_sdk_s3::{ByteStream, Client, Error, Region};
-use std::path::Path;
-use std::str;
+use aws_sdk_s3::{Client, Error, Region};
 use uuid::Uuid;
 
+// snippet-end:[rust.example_code.s3.basics.import]
+
+// snippet-start:[rust.example_code.s3.basics.main]
 #[tokio::main]
 async fn main() -> Result<(), Error> {
     let (region, client, bucket_name, file_name, key, target_key) = initialize_variables().await;
@@ -24,7 +25,9 @@ async fn main() -> Result<(), Error> {
 
     Ok(())
 }
+// snippet-end:[rust.example_code.s3.basics.main]
 
+// snippet-start:[rust.example_code.s3.basics.initialize_variables]
 async fn initialize_variables() -> (Region, Client, String, String, String, String) {
     let region_provider = RegionProviderChain::first_try(Region::new("us-west-2"));
     let region = region_provider.region().await.unwrap();
@@ -39,7 +42,9 @@ async fn initialize_variables() -> (Region, Client, String, String, String, Stri
 
     (region, client, bucket_name, file_name, key, target_key)
 }
+// snippet-end:[rust.example_code.initialize_variables]
 
+// snippet-start:[rust.example_code.s3.basics.run_s3_operations]
 async fn run_s3_operations(
     region: Region,
     client: Client,
@@ -48,127 +53,16 @@ async fn run_s3_operations(
     key: String,
     target_key: String,
 ) -> Result<(), Error> {
-    create_bucket(&client, &bucket_name, region.as_ref()).await?;
-    upload_object(&client, &bucket_name, &file_name, &key).await?;
-    download_object(&client, &bucket_name, &key).await?;
-    copy_object(&client, &bucket_name, &key, &target_key).await?;
-    list_objects(&client, &bucket_name).await?;
-    delete_objects(&client, &bucket_name).await?;
-    delete_bucket(&client, &bucket_name).await?;
+    s3_service::create_bucket(&client, &bucket_name, region.as_ref()).await?;
+    s3_service::upload_object(&client, &bucket_name, &file_name, &key).await?;
+    s3_service::download_object(&client, &bucket_name, &key).await?;
+    s3_service::copy_object(&client, &bucket_name, &key, &target_key).await?;
+    s3_service::list_objects(&client, &bucket_name).await?;
+    s3_service::delete_objects(&client, &bucket_name).await?;
+    s3_service::delete_bucket(&client, &bucket_name).await?;
 
     Ok(())
 }
+// snippet-end:[rust.example_code.s3.basics.run_s3_operations]
 
-async fn delete_bucket(client: &Client, bucket_name: &str) -> Result<(), Error> {
-    client.delete_bucket().bucket(bucket_name).send().await?;
-    println!("bucket deleted");
-    Ok(())
-}
-
-async fn delete_objects(client: &Client, bucket_name: &str) -> Result<(), Error> {
-    let objects = client.list_objects_v2().bucket(bucket_name).send().await?;
-
-    let mut delete_objects: Vec<ObjectIdentifier> = vec![];
-    for obj in objects.contents().unwrap_or_default() {
-        let obj_id = ObjectIdentifier::builder()
-            .set_key(Some(obj.key().unwrap().to_string()))
-            .build();
-        delete_objects.push(obj_id);
-    }
-    client
-        .delete_objects()
-        .bucket(bucket_name)
-        .delete(Delete::builder().set_objects(Some(delete_objects)).build())
-        .send()
-        .await?;
-
-    let objects: ListObjectsV2Output = client.list_objects_v2().bucket(bucket_name).send().await?;
-    match objects.key_count {
-        0 => Ok(()),
-        _ => Err(Error::Unhandled(Box::from(
-            "There were still objects left in the bucket.",
-        ))),
-    }
-}
-
-async fn list_objects(client: &Client, bucket_name: &str) -> Result<(), Error> {
-    let objects = client.list_objects_v2().bucket(bucket_name).send().await?;
-    println!("Objects in bucket:");
-    for obj in objects.contents().unwrap_or_default() {
-        println!("{:?}", obj.key().unwrap());
-    }
-
-    Ok(())
-}
-
-async fn copy_object(
-    client: &Client,
-    bucket_name: &str,
-    object_key: &str,
-    target_key: &str,
-) -> Result<(), Error> {
-    let mut source_bucket_and_object: String = "".to_owned();
-    source_bucket_and_object.push_str(bucket_name);
-    source_bucket_and_object.push('/');
-    source_bucket_and_object.push_str(object_key);
-
-    client
-        .copy_object()
-        .copy_source(source_bucket_and_object)
-        .bucket(bucket_name)
-        .key(target_key)
-        .send()
-        .await?;
-
-    Ok(())
-}
-
-async fn download_object(client: &Client, bucket_name: &str, key: &str) -> Result<(), Error> {
-    let resp = client
-        .get_object()
-        .bucket(bucket_name)
-        .key(key)
-        .send()
-        .await?;
-    let data = resp.body.collect().await;
-    println!(
-        "Data from downloaded object: {:?}",
-        data.unwrap().into_bytes().slice(0..20)
-    );
-
-    Ok(())
-}
-
-async fn upload_object(
-    client: &Client,
-    bucket_name: &str,
-    file_name: &str,
-    key: &str,
-) -> Result<(), Error> {
-    let body = ByteStream::from_path(Path::new(file_name)).await;
-    client
-        .put_object()
-        .bucket(bucket_name)
-        .key(key)
-        .body(body.unwrap())
-        .send()
-        .await?;
-
-    println!("Uploaded file: {}", file_name);
-    Ok(())
-}
-
-async fn create_bucket(client: &Client, bucket_name: &str, region: &str) -> Result<(), Error> {
-    let constraint = BucketLocationConstraint::from(region);
-    let cfg = CreateBucketConfiguration::builder()
-        .location_constraint(constraint)
-        .build();
-    client
-        .create_bucket()
-        .create_bucket_configuration(cfg)
-        .bucket(bucket_name)
-        .send()
-        .await?;
-    println!("{}", bucket_name);
-    Ok(())
-}
+// snippet-end:[rust.example_code.s3.scenario_getting_started.bin]
