+++ rust_dev_preview/s3/src/bin/s3-helloworld.rs
@@ -4,8 +4,7 @@
  */
 
 use aws_config::meta::region::RegionProviderChain;
-use aws_sdk_s3::ByteStream;
-use aws_sdk_s3::{Client, Error, Region, PKG_VERSION};
+use aws_sdk_s3::{ByteStream, Client, Error, Region, PKG_VERSION};
 use std::path::Path;
 use std::process;
 use structopt::StructOpt;
